refactor(cli): migrate headless/subagent run paths to internal/cli/headless (#363)
Move headless_session.go, subagent_rpc.go and their tests into
internal/cli/headless and extract dispatch's headless run block into
headless.Run(RunParams). Export SessionStore/PrintSessions/
MostRecentSessionID/RunSubAgentRPC/ParseOutputMode; keep the run
lifecycle helpers unexported. cmd/pigo/dispatch now resolves Mode+Env
and delegates; interactive.go's sessionStore() is a thin wrapper over
headless.SessionStore().
